Fix Stepper first label not shown bug
The label for first step depends on currently selected license: it's set to Attribution if it's a 'BY' license and 'Public Domain' if 'CC0' is selected. When the label was based on computed property translated by i18n, the label didn't change. So, the change was moved to a method.

Signed-off-by: Olga Bulat <obulat@gmail.com>