Fix layout
Footer 'Donate' button did not fit in a 1/4-sized column on tablet-sized layout, so the the app had a white margin on the right. Making all columns the same size on tablet size fixed this issue. Navbar menu burger button also had a gray background which was set to white.